requirements.txt
huggingface_hub
datasets
torch
torchvision
tqdm
pillow